docker/helios/src/requirements.txt
Django==2.2.3
django_webtest>=1.9
psycopg2==2.8.3
bleach==3.1.0
boto==2.27.0 
validate_email==1.3
oauth2client==4.1.3
rollbar==0.12.1
pycrypto==2.6.1
celery==4.4.7
python3-openid==3.1.0
html5lib==1.0.1 # Dependency of bleach

dj_database_url==0.5.0
django-auth-ldap==2.1.0

python-keycloak-client==0.2.2

pluginlib==0.8.2
drest==0.9.12

django-widget-tweaks==1.4.5

# django-session-timeout==0.1.0

gunicorn==20.0.4

# anyjson==0.3.3  ?? not imported by anything
# pyparsing==2.4.0  ?? not imported by anything
# python-dateutil>=1.5  ?? not imported by anything
# gunicorn==19.9
# requests==2.21.0
# unicodecsv==0.14.1
# dj_database_url==0.3.0
# django-sslify==0.2.7 # sslify.middleware.SSLifyMiddleware is commented out in settings.py
# webtest==2.0.18
# django-secure==1.0.1 # ?? Merged into django 1.8
# django-ses==0.6.0  # For Amazon's Simple Email Service



